+++ backend/models/Player.cs
@@ -4,7 +4,7 @@ using System.ComponentModel.DataAnnotations;
 public class Player
 {
     [Key]
-    public int Id { get; set; }
+    public short Id { get; set; }
     public string Username { get; set; } = string.Empty;
     public int Coins { get; set; }
     public int Skill { get; set; }
@@ -12,7 +12,7 @@ public class Player
     // Empty constructor is often helpful for frameworks
     public Player() { }
 
-    public Player(int id, string username, int coins, int skill)
+    public Player(short id, string username, int coins, int skill)
     {
         Id = id;
         Username = username;
